chore: component 폴더명 components로 변경
