+++ apps/commerce-client-web/src/app/(auth)/join/components/completion-step.tsx
