+++ apps/commerce-client-web/src/app/(auth)/join/components/step-indicator.tsx
