+++ apps/commerce-client-web/src/app/(auth)/join/components/terms-agreement.tsx
