+++ apps/commerce-client-web/src/app/(default)/(layout)/components/header-button.tsx
@@ -1,22 +1,23 @@
 import { Button } from '@/components/ui/button';
 import Link from 'next/link';
 import { ShoppingCart } from 'lucide-react';
+import useAuthStore from '@/stores/use-auth-store';
 
 const HeaderButton = () => {
-  const isLogin = false; // 현재는 로그인 상태를 false로 설정
+  const { isAuthenticated, logout } = useAuthStore();
+
   return (
     <>
       <Link href={'/cart'}>
         <ShoppingCart />
       </Link>
-      {isLogin ? (
+      {isAuthenticated ? (
         <>
-          <Button onClick={logout}>로그아웃</Button>
-          <Button asChild>
-            <Link href={'/myPage'}>마이페이지</Link>
+          <Button variant={'outline'} onClick={logout}>
+            로그아웃
           </Button>
           <Button asChild>
-            <Link href={'/myPage'}>장바구니</Link>
+            <Link href={'/myPage'}>마이페이지</Link>
           </Button>
         </>
       ) : (
